Add the backend to the cli function.
The methods uses Balena API to filter devices and enviroments variables.
It validates the input values to, raising exceptions if some of the
inputs is not valid (environment model invalid, environment variable not
existant, value to filter without matches).

It will also need one more level of fitering for the services model (app
services and device services) because it does not allow to check for
environment variables for a specific service name.