Refactor to interact with the reviewed releaser.
We needed a major overhauling over the releaser, so, naturally, the cli
has to be adapted to work with the new functions. Now devices returns a
list of dict with device_uuid:envar_key searched, so its easier to pass
the values to the remover function and we save a second call to the
filtering function.