Major refactor to add service filtering.
The overhaul of the releaser serves to add service filtering for envars,
as well to make the code more simple and concise. Now, the devices are
returned with the envar_id of the envar that makes then fulfill the
filtering condition. This helps us to later remove the envar,
but will also become handy if we eventually would want to update the envars
or do another operations. Also, its easier to get from the SDK the info
necessary to compare and filter (Sadly, Balena SDK suffers of a annoying lack
of abstractions and requieres a lot of 'data massaging' to get the
filtering done right).